+++ packages/nextjs-testapp/.wundergraph/wundergraph.server.ts
@@ -1,6 +1,6 @@
 import { GraphQLEnumType, GraphQLList, GraphQLNonNull, GraphQLObjectType, GraphQLSchema, GraphQLString } from 'graphql';
 import { configureWunderGraphServer } from '@wundergraph/sdk';
-import type { HooksConfig } from './generated/wundergraph.hooks';
+import type { AuthenticationResponse, HooksConfig } from './generated/wundergraph.hooks';
 import type { InternalClient } from './generated/wundergraph.internal.client';
 
 const testEnum = new GraphQLEnumType({
@@ -15,11 +15,31 @@ const testEnum = new GraphQLEnumType({
 	},
 });
 
-export default configureWunderGraphServer<HooksConfig, InternalClient>((serverContext) => ({
+export default configureWunderGraphServer<HooksConfig, InternalClient>(() => ({
 	hooks: {
+		authentication: {
+			async revalidate(): Promise<AuthenticationResponse> {
+				return {
+					status: 'ok',
+					user: {
+						name: 'test',
+					},
+				};
+			},
+
+			async postAuthentication(hook) {},
+			async mutatingPostAuthentication(hook): Promise<AuthenticationResponse> {
+				return {
+					status: 'ok',
+					user: {
+						name: 'test',
+					},
+				};
+			},
+		},
 		queries: {
 			FakeWeather: {
-				mockResolve: async (hookContext) => {
+				mockResolve: async (hook) => {
 					return {
 						data: {
 							getCityByName: {