+++ pkg/apihandler/apitransport.go
@@ -9,14 +9,13 @@ import (
 	"net/http"
 	"net/http/httputil"
 	"net/url"
-	"strings"
 	"time"
 
 	"github.com/buger/jsonparser"
 	"github.com/dgrijalva/jwt-go"
 	"github.com/wundergraph/graphql-go-tools/pkg/pool"
+	"github.com/wundergraph/wundergraph/pkg/hooks"
 	"github.com/wundergraph/wundergraph/pkg/loadvariable"
-	"github.com/wundergraph/wundergraph/pkg/middlewareclient"
 	pool2 "github.com/wundergraph/wundergraph/pkg/pool"
 	"github.com/wundergraph/wundergraph/types/go/wgpb"
 
@@ -30,10 +29,10 @@ type ApiTransport struct {
 	upstreamAuthConfigurations map[string]*wgpb.UpstreamAuthentication
 	onRequestHook              map[string]struct{}
 	onResponseHook             map[string]struct{}
-	hooksClient                *middlewareclient.MiddlewareClient
+	hooksClient                *hooks.Client
 }
 
-func NewApiTransport(tripper http.RoundTripper, api *wgpb.Api, hooksClient *middlewareclient.MiddlewareClient, enableDebugMode bool) http.RoundTripper {
+func NewApiTransport(tripper http.RoundTripper, api *wgpb.Api, hooksClient *hooks.Client, enableDebugMode bool) http.RoundTripper {
 	transport := &ApiTransport{
 		roundTripper:               tripper,
 		debugMode:                  enableDebugMode,
@@ -199,11 +198,11 @@ func (t *ApiTransport) internalGraphQLRoundTrip(request *http.Request) (res *htt
 	}
 
 	if clientRequest, ok := request.Context().Value(pool2.ClientRequestKey).(*http.Request); ok {
-		requestJSON, err := HttpRequestToWunderGraphRequestJSON(clientRequest, false)
+		requestJSON, err := hooks.HttpRequestToWunderGraphRequestJSON(clientRequest, false)
 		if err != nil {
 			return nil, err
 		}
-		requestBody, err = jsonparser.Set(requestBody, requestJSON, "__wg", "client_request")
+		requestBody, err = jsonparser.Set(requestBody, requestJSON, "__wg", "clientRequest")
 	}
 
 	req, err := http.NewRequestWithContext(request.Context(), request.Method, request.URL.String(), ioutil.NopCloser(bytes.NewReader(requestBody)))
@@ -214,25 +213,6 @@ func (t *ApiTransport) internalGraphQLRoundTrip(request *http.Request) (res *htt
 	req.Header = request.Header.Clone()
 
 	return t.roundTripper.RoundTrip(req)
-}
-
-type WunderGraphRequest struct {
-	Method     string            `json:"method"`
-	RequestURI string            `json:"requestURI"`
-	Headers    map[string]string `json:"headers"`
-	Body       json.RawMessage   `json:"body,omitempty"`
-}
-
-type OnRequestHookPayload struct {
-	Request       WunderGraphRequest `json:"request"`
-	OperationName string             `json:"operationName"`
-	OperationType string             `json:"operationType"`
-}
-
-type OnRequestHookResponse struct {
-	Skip    bool                `json:"skip"`
-	Cancel  bool                `json:"cancel"`
-	Request *WunderGraphRequest `json:"request"`
 }
 
 func (t *ApiTransport) handleOnRequestHook(r *http.Request, metaData *OperationMetaData) (*http.Request, error) {
@@ -246,11 +226,11 @@ func (t *ApiTransport) handleOnRequestHook(r *http.Request, metaData *OperationM
 			return nil, err
 		}
 	}
-	payload := OnRequestHookPayload{
-		Request: WunderGraphRequest{
+	payload := hooks.OnRequestHookPayload{
+		Request: hooks.WunderGraphRequest{
 			Method:     r.Method,
 			RequestURI: r.URL.String(),
-			Headers:    headerSliceToCSV(r.Header),
+			Headers:    hooks.HeaderSliceToCSV(r.Header),
 			Body:       body,
 		},
 		OperationName: metaData.OperationName,
@@ -265,11 +245,13 @@ func (t *ApiTransport) handleOnRequestHook(r *http.Request, metaData *OperationM
 			hookData, _ = jsonparser.Set(hookData, userJson, "user")
 		}
 	}
-	out, err := t.hooksClient.DoGlobalRequest(r.Context(), middlewareclient.HttpTransportOnRequest, hookData)
+
+	out, err := t.hooksClient.DoGlobalRequest(r.Context(), hooks.HttpTransportOnRequest, hookData)
 	if err != nil {
 		return nil, err
 	}
-	var response OnRequestHookResponse
+
+	var response hooks.OnRequestHookResponse
 	err = json.Unmarshal(out.Response, &response)
 	if err != nil {
 		return nil, err
@@ -286,31 +268,10 @@ func (t *ApiTransport) handleOnRequestHook(r *http.Request, metaData *OperationM
 		if err != nil {
 			return nil, err
 		}
-		req.Header = headerCSVToSlice(response.Request.Headers)
+		req.Header = hooks.HeaderCSVToSlice(response.Request.Headers)
 		return req, nil
 	}
 	return r, nil
-}
-
-type WunderGraphResponse struct {
-	StatusCode int               `json:"statusCode"`
-	Status     string            `json:"status"`
-	Method     string            `json:"method"`
-	RequestURI string            `json:"requestURI"`
-	Headers    map[string]string `json:"headers"`
-	Body       json.RawMessage   `json:"body,omitempty"`
-}
-
-type OnResponseHookPayload struct {
-	Response      WunderGraphResponse `json:"response"`
-	OperationName string              `json:"operationName"`
-	OperationType string              `json:"operationType"`
-}
-
-type OnResponseHookResponse struct {
-	Skip     bool                 `json:"skip"`
-	Cancel   bool                 `json:"cancel"`
-	Response *WunderGraphResponse `json:"response"`
 }
 
 func (t *ApiTransport) handleOnResponseHook(r *http.Response, metaData *OperationMetaData) (*http.Response, error) {
@@ -318,11 +279,11 @@ func (t *ApiTransport) handleOnResponseHook(r *http.Response, metaData *Operatio
 	if err != nil {
 		return nil, err
 	}
-	payload := OnResponseHookPayload{
-		Response: WunderGraphResponse{
+	payload := hooks.OnResponseHookPayload{
+		Response: hooks.WunderGraphResponse{
 			Method:     r.Request.Method,
 			RequestURI: r.Request.URL.String(),
-			Headers:    headerSliceToCSV(r.Header),
+			Headers:    hooks.HeaderSliceToCSV(r.Header),
 			Body:       body,
 			StatusCode: r.StatusCode,
 			Status:     r.Status,
@@ -339,11 +300,13 @@ func (t *ApiTransport) handleOnResponseHook(r *http.Response, metaData *Operatio
 			hookData, _ = jsonparser.Set(hookData, userJson, "user")
 		}
 	}
-	out, err := t.hooksClient.DoGlobalRequest(r.Request.Context(), middlewareclient.HttpTransportOnResponse, hookData)
+
+	out, err := t.hooksClient.DoGlobalRequest(r.Request.Context(), hooks.HttpTransportOnResponse, hookData)
 	if err != nil {
 		return nil, err
 	}
-	var response OnResponseHookResponse
+
+	var response hooks.OnResponseHookResponse
 	err = json.Unmarshal(out.Response, &response)
 	if err != nil {
 		return nil, err
@@ -358,7 +321,7 @@ func (t *ApiTransport) handleOnResponseHook(r *http.Response, metaData *Operatio
 	if response.Response != nil {
 		r.StatusCode = response.Response.StatusCode
 		r.Body = ioutil.NopCloser(bytes.NewBuffer(response.Response.Body))
-		r.Header = headerCSVToSlice(response.Response.Headers)
+		r.Header = hooks.HeaderCSVToSlice(response.Response.Headers)
 	}
 	return r, nil
 }
@@ -406,34 +369,4 @@ func (t *ApiTransport) handleUpstreamAuthentication(request *http.Request, auth 
 		return fmt.Errorf("not implemented")
 	}
 	return nil
-}
-
-func HttpRequestToWunderGraphRequestJSON(r *http.Request, withBody bool) ([]byte, error) {
-	var body []byte
-	if withBody {
-		body, _ = ioutil.ReadAll(r.Body)
-		r.Body = ioutil.NopCloser(bytes.NewBuffer(body))
-	}
-	return json.Marshal(WunderGraphRequest{
-		Method:     r.Method,
-		RequestURI: r.URL.String(),
-		Headers:    headerSliceToCSV(r.Header),
-		Body:       body,
-	})
-}
-
-func headerSliceToCSV(headers map[string][]string) map[string]string {
-	result := make(map[string]string, len(headers))
-	for k, v := range headers {
-		result[k] = strings.Join(v, ",")
-	}
-	return result
-}
-
-func headerCSVToSlice(headers map[string]string) map[string][]string {
-	result := make(map[string][]string, len(headers))
-	for k, v := range headers {
-		result[k] = []string{v}
-	}
-	return result
 }